+++ src/main/java/dts/logic/EnhancedOperationsServiceImplementation.java
@@ -2,6 +2,7 @@ package dts.logic;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -12,10 +13,22 @@ import org.springframework.transaction.annotation.Transactional;
 
 import dts.converter.OperationConverter;
 import dts.dao.IdGeneratorDao;
+import dts.dao.ItemDao;
 import dts.dao.OperationDao;
+import dts.dao.UserDao;
 import dts.data.IdGeneratorEntity;
+import dts.data.ItemEntity;
 import dts.data.OperationEntity;
+import dts.data.UserEntity;
+import dts.data.UserRole;
+import exceptions.AdminNotFoundException;
+import exceptions.InvalidOperationTypeException;
+import exceptions.InvalidUserException;
+import exceptions.ItemNotFoundException;
+import exceptions.UserNotFoundException;
+import models.operations.ItemId;
 import models.operations.OperationId;
+import models.users.UserId;
 import boundaries.OperationBoundary;
 
 @Service
@@ -25,6 +38,8 @@ public class EnhancedOperationsServiceImplementation implements OperationsServic
 	private OperationDao operationDao;
 	private OperationConverter operationConverter;
 	private IdGeneratorDao idGeneratorDao;
+	private UserDao userDao;
+	private ItemDao itemDao;
 
 	@Value("${spring.application.name:default_space_name}")
 	public void setSpaceName(String spaceName) {
@@ -32,17 +47,41 @@ public class EnhancedOperationsServiceImplementation implements OperationsServic
 	}
 
 	@Autowired
-	public EnhancedOperationsServiceImplementation(OperationDao operationDao, OperationConverter operationConverter,
-			IdGeneratorDao idGeneratorDao) {
+	public EnhancedOperationsServiceImplementation(OperationDao operationDao, 
+			OperationConverter operationConverter,
+			IdGeneratorDao idGeneratorDao,
+			UserDao userDao,
+			ItemDao itemDao) {
 		this.operationDao = operationDao;
 		this.operationConverter = operationConverter;
 		this.idGeneratorDao = idGeneratorDao;
+		this.userDao = userDao;
+		this.itemDao = itemDao;
 	}
 
 	@Override
 	@Transactional
 	public Object invokeOpreation(OperationBoundary operation) {
 		OperationEntity operationEntity = this.operationConverter.toEntity(operation);
+		if(operationEntity.getType().isEmpty() ||
+		   operationEntity.getType() == null)
+		{
+			throw new InvalidOperationTypeException("Invalid type: " + 
+					operationEntity.getType() + " for operation: " + 
+					operationEntity.getOperationId());
+		}
+		Optional<UserEntity> existingUser = this.userDao.findById(operationEntity.getInvokedBy());
+		if (!existingUser.isPresent())
+		{			
+			throw new UserNotFoundException("user with email: " + operationEntity.getInvokedBy() + "does not exist");
+		}
+		Optional<ItemEntity> existingItem = this.itemDao.findById(operationEntity.getItem());
+		if (!existingItem.isPresent())
+		{			
+			throw new ItemNotFoundException("Item with id: " + operationEntity.getItem() + 
+					" does not exist");
+		}
+
 		IdGeneratorEntity idGeneratorEntity = new IdGeneratorEntity();
 		idGeneratorEntity = this.idGeneratorDao.save(idGeneratorEntity);
 		Long newId = idGeneratorEntity.getId();
@@ -57,6 +96,17 @@ public class EnhancedOperationsServiceImplementation implements OperationsServic
 	@Override
 	@Transactional(readOnly = true)
 	public List<OperationBoundary> getAllOperations(String adminSpace, String adminEmail) {
+		Optional<UserEntity> existingAdmin = this.userDao.findById(new UserId(adminSpace, adminEmail).toString());
+		if(!existingAdmin.isPresent())
+		{
+			throw new AdminNotFoundException("Admin with email: " + existingAdmin + "does not exist");
+		}
+		UserEntity existingAdminEntity = existingAdmin.get();
+		if(!existingAdminEntity.getRole().equals(UserRole.ADMIN))
+		{
+			throw new InvalidUserException("Invalid role: " + existingAdminEntity.getRole()
+			+ " for user: " + existingAdminEntity.getUsername());
+		}
 		return StreamSupport
 				.stream(this.operationDao.findAll().spliterator(), false) // Iterable to Stream<OperationEntity>,
 				.map(entity -> this.operationConverter.toBoundary(entity)) // Stream<OperationBoundary>
@@ -66,6 +116,17 @@ public class EnhancedOperationsServiceImplementation implements OperationsServic
 	@Override
 	@Transactional
 	public void deleteAllActions(String adminSpace, String adminEmail) {
+		Optional<UserEntity> existingAdmin = this.userDao.findById(new UserId(adminSpace, adminEmail).toString());
+		if(!existingAdmin.isPresent())
+		{
+			throw new AdminNotFoundException("Admin with email: " + existingAdmin + "does not exist");
+		}
+		UserEntity existingAdminEntity = existingAdmin.get();
+		if(!existingAdminEntity.getRole().equals(UserRole.ADMIN))
+		{
+			throw new InvalidUserException("Invalid role: " + existingAdminEntity.getRole()
+			+ " for user: " + existingAdminEntity.getUsername());
+		}
 		this.operationDao.deleteAll();
 	}
